feat: add crate feature log-lock_events
This feature logs lock events: try_acquire, acquire, release which is
useful for debugging lock problems such as a lock that is never
released.

It already existed in the codebase but was disabled in 'if false'
blocks.

It is now made into a crate feature, so it is more discoverable and
easier to enable.