feat: include location in log-lock-events feature
When looking at logs to debug lock issues it is helpful/necessary
to see where the lock-acquisition [attempt] happened.

To that end, it was necessary to do some refactoring.  The end result
is a bit cleaner and more consistent than before.

Now AtomicRw/AtomicMutex only track acquisition location and instant,
and only if features track-lock-location and track-lock-time are
enabled.

The callback handler in application code (lib.rs) receives the
location and instant in the LockEvent and logs according to app
needs/heuristics.