refactor: Reduce visibility of CLI args on `GlobalState`
CLI arguments can never change, so they need not be hidden behind a
smart pointer. So instead of (a) acquiring the lock on `GlobalState`
and then (b) getting the CLI arguments from there, it is better to
clone the CLI arguments on `GlobalStateLock` (which is where the smart
pointer lives).