perf: disable cpu temp in dashboard data.
closes #244

We (temporarily?) disable the call to cpu_temp_inner() inside
the dashboard_overview RPC.

The call performs blocking IO reading OS files and usually
takes at least .008 secs, which is way too long for async code.

It is arguable this does not belong in neptune-core at all.
neptune-dashboard is designed to run on the same machine, so it
could just call the cpu_temp api itself.  Plenty of (better)
solutions exist for remote node monitoring.  This feels like
its outside the scope of neptune-core's mission.