perf: avoid extra tip5 hash in Utxo::hash()
previously the 'impl std::hash::Hash for Utxo' was performing
a tip5 hash as input to std::hash::Hash::hash().

In other words, double-hashing.

this commit removes the tip5 hash, and simply uses self.encode()
(BFieldCode::encode()) as the input.

Note that this is mainly for purpose of storing Utxo in a HashMap.
It does not affect the publicly visible hash/id.