perf: avoid extra tip5 hash in TypeScript Hash impl
previously the 'impl std::hash::Hash for TypeScript' was performing
a tip5 hash as input to std::hash::Hash::hash().

In other words, double-hashing.

this commit removes the tip5 hash, and simply uses self.encode()
(BFieldCode::encode()) as the input.

Note that the Hash impl is mainly for purpose of filtering dups.