refactor: make TransactionKernel immutable
This refactor was done in order to cache the result of
MastHash::mast_sequences(). It turns out the cache is not much
of a measurable win so far.  However...

In order to include a OnceLock in TransactionKernel it was necessary
to make the struct immutable. The resulting code is worth keeping
since it forces creation of a new TransactionKernel any time
a field is altered, which makes such changes much more explicit and
easily searchable.

changes:
 * add #[readonly:make] annotation to TransactionKernel
 * add mast_sequences OnceLock cache to TransactionKernel
 * impl PartialEq for TransactionKernel, to ignore mast_sequences
 * add TransactionKernelProxy with pub fields for creation and
    destructuring of TransactionKernel
 * add TransactionKernelModifier (builder pattern) for generating modified
   copies of a TransactionKernel
 * update code/tests to use the above.