fix: Check emptiness, not absence, of SWBFI chunks
Test
`update_mutator_set_rollback_many_blocks_multiple_inputs_outputs_test`
was asserting that the SWBFI MMR of the mutator set following the
genesis block was empty. But it is not empty if there are enough
premine UTXOs to induce a window slide. The SWBFI is empty, but with
window slides its according MMR will have leafs representing empty
chunks. The solution is instead to iterate over all chunks and
verify that they are empty.