archival-mutator-set: Read all chunks in one go
Had to replace a stream_many with a get_many in ArchivalMutatorSet
because of a rustc type limitation. I prefer this approach though,
as the sequence in question is only 45 elements long.

Without this change, the RPC tasks could not be spawned in lib.rs,
because of life time errors.