chore: remove PartialEq - block_primitive_witness
block_primitive_witness contains a OnceLock field 'maybe_body' for cacheing
a computed result.  As such it should not derive PartialEq because that can
cause the equality to fail depending on the field's state.

It turns out that the PartialEq, Eq were not being used anywhere so they
are now removed.  If ever needed, PartialEq can be impl'd manually.

Since maybe_body depends on other fields in the struct, the struct must
be immutable outside the module.  To that end all fields are made
private and a getter is added.