claim_utxo: verify sender_randomness equality in wallet check
When checking if a UTXO/offchain UTXO notification has already been
claimed by this client's wallet, verify that not just the UTXO matches
an entry in the wallet, but that the sender randomness also agrees.

It's quite conceivable that a party will send the same transaction to
an address multiple times, and we want to ensure that the wallet can
catch all such UTXOs.