claim_utxo: Remove option to read ciphertext from stdin
This option was very cool but due to a stupid max length of lines in the
terminal, enforced in the Linux kernel, of 4096 characters, it doesn't
work. Our ciphertexts are ~4500 characters long when they're
Bech32-encoded. We could use another encoding to get around that
problem, I guess. But for now, it's best just to remove this option.

Cf.
https://github.com/torvalds/linux/blame/f40ddce88593482919761f74910f42f4b84c004b/drivers/tty/n_tty.c#L1681

Co-authored-by: Alan Szepieniec <alan@neptune.cash>