fix: send no longer pause/unpause mining/composing.
closes #253.

It is unnecessary to pause/resume mining during send because send
is now quite fast so there is no CPU contention with the mining task.

Worse, pausing causes any composing work already done on current block
to be aborted and lost.

This commit simply removes the pause/resume functionality from
the send rpcs.