test(Block): Add test that double-spending blocks are rejected
Includes a rewrite of helper function for constructing a new block that
makes it easier to create a new block from a BlockPrimitiveWitness.
Specifically, a new function `block_template_from_primitive_witness` is
added which is used in this test as well as in the call graph for
creating a new block from a transaction and a predecessor.

Also adds an "unsafe" version of MutatorSetUpdate's "apply" function
that allows the caller to calculate a new MS accumulator which ignores
double spends. This was needed to allow the BlockPrimitiveWitness
constructor to complete the construction of a block with a double-
spending transaction.

TODO in future commit: Add a test that verifies that double spend
*within* one block is also disallowed.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>