style: Prefer `display()` over `to_string_lossy()`
For printing directory and file paths, `display()` is preferable over
`to_string_lossy()`. Both deal with unicode characters in a safe way,
but `to_string_lossy()` replaces all of them with the replacement
character `�`, resulting in loss of information and making it harder
to find (or create) the file in question.