refactor(ConsensusProgram): Take owned argument
Honestly declare ownership requirements of `ConsensusProgram::prove`.

Previously, the interface lied: it suggested that only a reference to
the Claim was required. In reality, an owned Claim was required, which
was produced through cloning.

Also, hide the interface in the documentation, since it is only needed
for benchmarks.

Also, remove superfluous block, decreasing nesting depth.