feat(peer-loop): Ask peer for connections first
Whenever we establish a connection with a new peer, the first thing to
do after the handshake is now to ask the peer for its list of peers.
This change is motivated by topological robustness in the context of
adversaries who deploy sybil or other attacks. In such a context, the
connection is not guaranteed to last and so we want to prioritize the
information that affects network topology -- blockchain data may be
gathered from other nodes.