fix(mine_loop): Avoid using different blocks in block composition (#267)
Prior to this fix, a race condition could lead to two different blocks
being used in the witness composition for a new block. This happened
since one block was input to the function, and inside the function to
construct the coinbase transaction, a block was read from global state.
With this fix, the block used as input argument is always used.

This closes #266.