fix: Split `ConnectionStatus` into `Internal` and `Transfer`
The `Transfer` variant is sent over the wire. The `Internal` variant
is used for internal error or success codes. Commit
5559043aaae402a7cef817b05351d4c8078f41b9 expanded `ConnectionStatus`
but only for internal bookkeeping. However, as this object was sent
over the wire to other peers, peers running the old software did not
recognize the new encoding. This commit fixes that error. Concretely,
the encoding of `TransferConnectionStatus` is identical to that of
`ConnectionStatus` before the offending commit.