fix(wallet_state): Never store same UTXO recovery data twice
In the case of a reorganization, the wallet state update must take care
to never add the same UTXO twice. This was already correctly handled for
the list of `MonitoredUtxo` but not for the file containing recovery
data for the received UTXOs. With this commit, the recovery data is
stored under the same conditions that a MonitoredUtxo is added.

This closes #272.