feat(wallet_state): Ensure no duplicated values in UTXO-recovery file
At startup, the list of monitored UTXOs is compared against the list of
UTXO-recovery elements stored in a file in the wallet directory. This
list may not contain duplicates as evidenced by #272 and
f4e3d49aae9ba6f1b9a7158a66a434f89e55c7d7. This commit adds a check that
there are no duplicates in this list, and reports the AOCL leaf index if
an element occurs twice, such that the user can manually remove it.