feat!(archival_state): Return block adding AOCL leaf
Add a method to `ArchivalState` that takes an AOCL leaf index and
returns the digest of the block belonging to the canonical chain that
added this leaf to the AOCL MMR.

This functionality is added to improve the recovery experience of
`MonitoredUtxo`, such that recovered data also contains history, i.e.
information about how old this incoming UTXO is. This new method will be
used in `GlobalState::restore_monitored_utxos_from_recovery_data`.