feat(native-currency): Time-lock half of coinbase
The tokenomics of Neptune Cash stipulate that half of all mining
be time-locked for three years. This commit time-locks half of the
coinbase. Half of the guesser fee is already automatically
time-locked.

Specifically, this commit modifies the tasm code for consensus program
`NativeCurrency` to verify that across all outputs (and counting half
of the transaction fee), at least half of the coinbase is time-locked
for three years.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>