feat: display symmetric key hash to avoid leaking secret
adds ReceivingAddress::to_display_bech32m() which Symmetric key
impls to encode the privacy_preimage() as bech32m instead of the
key itself.

This is then used for display/id purposes in places such as the
dashboard "Addresses" screen.

This is experimental. It could have unforseen knock-on effects.

other changes:

* improve variable naming, comments
* add test known_keys_are_unique
* add test derivation_counter_persists_across_restart
* store known-keys in ordered Vec.