fix(mine_loop): Make hash rate independent of tx size
Prior to this commit, the hash rate was lower for blocks with many
inputs/outputs in their transaction. This was because the block body
MAST hash was not cached, so it was recalculated for each PoW iteration.

This strongly incentivized miners to mine empty or small blocks, to the
detriment of the ability to have transactions confirmed and merged.