perf!(SingleProof): Delete Merge and Update programs
Drop the consensus programs Merge and Update by merging their logic into
SingleProof.

This yields a speedup of 100 % for the update of a transaction to a new
mutator set, and a speedup of 50 % for the merging of two transactions!

This is achieved by taking what was previously the Update/Merge program
and invoking their logic in branches called from SingleProof. These
branches are implemented as snippets, using the BasicSnippet trait.
These new snippets are, however, not tested in the normal way (through
e.g. implementations of the `Procedure` trait) but instead tested with
the "source" testing framework that consensus programs normally use.
This means that the tests of the update/merge branches now live in
SingleProof instead of their respective modules.

No tests were harmed in the writing of this optimization. However, the
merge branch is missing negative test cases, as was also the case prior
to this commit. Cf. #275.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>