test: validate we can send to symmetric keys
refactors the rpc_server::send_to_many() test to operate once for
each KeyType variant. Each variant is used as the recipient's keytype.
This means it now tests sending to a symmetric key as well as a
generation key.  It will also test sending to any future keytype we
might add.