test: Add positive and negative tests for negative fees
In particular, this commit adds a new arbitrary implementation for
`PrimitiveWitness`; this one takes the transaction fee as argument.
This transaction fee can be negative and even (as far as the arbitrary
implementation is concerned) invalid.

The tests do not pass yet. Future changes to `NativeCurrency` and
`SingleProof` that accomodate negative fees will fix these failures.
The tests are there for test-driven development.