fix: Spawn task for block proof verification
Verifying a block proof takes on the order of hundreds of
microseconds. It is good practice to wrap such calls inside
of a `spawn_blocking` invocation, which is what this commit does.

In order to use `spawn_blocking`, the function in question must be
made `async`. This change percolates up the call graph.