fix: Spawn task for block proof verification
Verifying a block proof takes on the order of hundreds of microseconds.
It is good practice to wrap such calls inside of a `spawn_blocking`
invocation, which is what this commit does.

Besides observing good practice, it seems as though this change fixes
issue #282. The reported crash was not yet observed with this change
in place. However, we do not understand the relevant mechanics or
have an educated guess as to why this change causes the issue to go
away. For all we know, the issue might appear again even with this
change.

In order to use `spawn_blocking`, the function in question must be
made `async`. This change percolates up the call graph.

The call to `triton_vm::verify` that pinpoints issue #282, or more
precisely, the `spawn_blocking` wrapper that contains it, is sandwiched
in between log statements. So if the node hangs again, the node operator
will be able to pinpoint the cause.