chore: neptune-cli offline cmds accept network
closes #256

small refactor to neptune-cli so that online commands obtain network
from neptune-core via rpc and offline commands accept a network
arg and optional data-dir arg.  This prevents a network mis-match
between neptune-cli and neptune-core.

also renames own-receiving-address to next-receiving-address.

These are some of the changes from pr #185 that were missed in the
merge.