test: add tests to verify key/addr format and derivation
adds two tests that derives keys and addresses respectively
and compares each set against a corresponding set of hard-coded
keys or addresses.

The hard-coded sets were generated from the alphanet tag.

The tests will fail if the key or address format ever changes
or the derivation scheme, or the bech32m address encoding.

-----

these tests are cherry-picked from branch
alphanet-with-generation-derive-test.

The tests presently fail on master, so they are placed
on a branch in case anyone wants to review them.