fix: Temporarily exceed max num peers when bootstrapping
There is a back-stop in place to ensure that the maximum number of
peers can never be exceeded, even accounting for race conditions.
However, the behavior of bootstrapper nodes is to disconnect from
the longest-lived node whenever the maximum is reached. In this case,
control might reach the backstop before the instruction to disconnect
has been processed.

The solution is to disable the backstop whenever a corresponding
disconnect-from-longest-lived-peer message has been spawned -- that
is, ignore the backstop when the node is configured to bootstrap with
the CLI argument `--bootstrap`.

@elielnfinic's analysis was invaluable for tracking down the root
cause of this bug.

Fixes #287.

@Co-authored-by: @elielnfinic <46634774+elielnfinic@users.noreply.github.com>