log(connect_to_peers): reduce log-noise on failed outgoing connections
Previously, the `error` log-level was used when outgoing connections
failed. With this commit, it is either `warn` or `info`. `warn` if the
user has explicitly requested to be connected to this peer in the CLI
arguments. And `info` if they have not, which would indicate this
connection is being made through peer-discovery.