test: Mimic real blocks better in mock_block generator
Instead of creating a fake block with only one composer-output, create
one with *two* outputs to better mimic the behavior of the actual
blocks. Also makes this `make_mock_block` generator async since it can
then use more of the real pipeline with which new blocks and coinbase
transactions are created.

The change in behavior in this widely used function forced a big diff
that, however, is contained to test-code.