fix(wallet_state): Avoid adding duplicate MUTXOs from different blocks
Instead of using the block hash to identify already-added monitored
UTXOs, we use the so-called "StrongUtxoKey" which contains both the
addition record and the AOCL leaf index. Only if these two values are
the same is the UTXO considered the same. Previously, we used block
hashes to check if two UTXOs are the same, but that did not catch the
corner-case where a composer receives two different PoW solutions to a
block proposal. In this case, the composer UTXOs should only be added
once, not twice as was previously the case.

This closes #309.