feat(archival_state): Add archival block MMR
Add a new field to the archival state, `archival_block_mmr` which is an
archival MMR of all block digests in the canonical chain.

This allows the client to instantly check if a block belongs to the canonical
chain, and thus to faster prune abandoned monitored UTXOs, which is the
motivation for this addition.

In the future, this data structure will allow archival nodes to prove
the canonicity of a block relative to a tip such that light nodes (that
only see the current tip) can verify that any received blocks belong to
the canonical chain.