feat(archival_state): Use block MMR to determine canonicity
Use the new field in archival state, archival block MMR, to determine if
a block belongs to the canonical chain. This replaces a potentially
expensive lookup on disk with a single lookup in an archival MMR. This
should make pruning of abandoned monitored UTXOs and responding to
`PeerMessage::BlockRequestBatch` *much* faster as they require
significantly less IO.

With this extra field in the archival state, it's probably prudent to do
a data integrity-check on startup though. That should be made a priority
ASAP.