fix(Block::is_valid): Fix false negative
The removal records were, prior to this commit, compared against the
mutator set field in the block body, not against the mutator set as it
looks after the guesser fees have been added. This lead to an error of
false negatives as the moving of the active window could invalidate
removal record membership proofs.