test(block): Fix negative test of block validity/bad block MMRA value
Fix a negative test that verifies that the block validity function fails
if the the `block_mmr_accumulator` value is not set correctly. Prior to
the mutations that invalidate the block, we verify that the block is
valid.

Also: This test only works because we verify the `block_mmr_accumulator`
prior to verifying the block proof, since the block proof is invalidated
by the new block_mmr_accumulator field value (since this modification
changes the block's claim). To address this problem (of not knowing what
you're testing in the negative tests), #293 must be solved sooner rather
than later.