test: Fix mempool test of reorganization/tx-updates
Actually create the new transaction proof that is the output of a update
process. This was previously not done, so the transactions became
invalid as the active window shifted. It was the
`is_confirmable_relative_to` method that failed prior to this commit,
not the `is_valid` as the latter does not take blockchain state into
account, only internal consistency.