refactor(single_proof): Disallow update of coinbase-transactions
This prevents a composer that wants to collect block subsidies from
reusing block transactions across blocks -- i.e. using the same block
transaction modulo an update across multiple blocks.

The reason a composer would want to do this is to compose faster as an
update is faster than the PrimitiveWitness -> ProofCollection ->
SingleProof ->(via merge) SingleProof pipeline. The reason we disallow
this is to remove an incentive to mine empty blocks and to remove an
incentive to reduce the anonymity-set by using repeated addition
records.

Anyone who wants to collect block subsidies will have to follow the long
pipeline shown above.