fix(`WalletState`): Receive premine at multiple addresses
We cannot be sure that premine recipients used key derivation
index 0 for their premine receiving address. Previously, if they
used a nonzero index for their premine receiving address, then
they would not receive their premine allocation. This commit fixes
that.

In particular, when updating the wallet state with the genesis block,
a bunch of receiving addresses are generated (not just the zeroth) and
the genesis block is scanned for any match.

Closes #300.