chore: add comment in Cargo.toml
adds a comment that some optional deps are duplicated
in dev-deps because they are required for unit testing.

there may be a way to do it that does not require duplication
but I haven't found it yet.

note that the duplication does not seem to be necessary with
nightly compiler.