test: Fix test of double-spend illegality
Fix a test verifying that a double-spend over two blocks is illegal and
does not pass the block validity check. This test was broken by the
update that disallowed the update of coinbase transactions, as the
previous implementation of this test reused a coinbase transaction
(with inputs -- so merged with a regular tx) for its update target.