feat(rpc): cookie based authentication
closes #326

adds a Token parameter to (almost) every RPC method.  The Token is an
enum presently with a single variant: Cookie.

Each time neptune-core starts a new valid Cookie is generated and stored
on disk in <root>/.cookie and in RAM.  local RPC clients with
permissions can read the .cookie file and include it with the RPC
request, thus proving they are authorized.

The primary benefit of this cookie system is that it is zero-conf, ie it
does not require the user to manually create a password.

To facilitate zero-conf, there is a new RPC api cookie_hint() which
returns the root data-dir and the network.  So even if neptune-core is
using a custom location for data-dir, a client can find it.

The cookie_hint() RPC can be disabled via flag --disable-cookie-hint. In
this case, client's fallback to using their own --data-dir flags.

Changes:

neptune-core:
* add --disable-cookie-hint to disable cookie_hint RPC.
  (prevent clients seeing data-dir path)
* impl Serialize, Deserialize for DataDirectory
* add DataDirectory::rpc_cookie_file_path()
* add data_dir to GlobalStateLock
* generate a new rpc cookie in <root>/.cookie each time neptune-core starts
* add rpc_auth module with Token, Cookie, auth errors.
* RPC:
    * keep a list of valid tokens in NeptuneRPCServer
    * add error module with enum RpcError.
    * add cookie_hint RPC
    * all methods require auth token and return RpcResult
      exceptions: network(), cookie_hint()
    * update rpc tests to use auth cookie token.

unit tests:
* new test: rpc_auth::token::cookie::auth()
* new test: rpc_auth::cookie::try_new_unique()
* new test: rpc_auth::cookie::auth()
* new test: rpc_auth::cookie::concurrency()

neptune-dashboard:
* query token_hint rpc, fallback to --data-dir or default.
* send auth Token with rpc requests

neptune-cli:
* query token_hint rpc, fallback to --data-dir or default.
* send auth Token with rpc requests
* make --data-dir a global option in neptune-cli so user can override
  automatic detection via token-hint.