style(`GlobalState`): Rename criterion function
`should_enter_into_sync_mode` -> `sync_mode_criterion`

The function evaluates a criterion. This criterion is the
antecedent. Entering into sync mode is the consequence, and not
a necessary one.

This particular antecedent can be evaluated and even satisfied
independently of the consequence being activated. The new name
communicates that while the consequence is linked, it does not
necessarily follow from the antecedent.