feat(`main_loop`): Tolerate arbitrarily deep reorganizations
Attach a new field `champion` to the `SyncAnchor` living on
`NetworkingState`, which records the block height and digest of the
block with largest height on the fork specified by the anchor.

When compiling a block batch request, use the digest given by this
champion. If the champion is not set, select block digests based on
a sparse collection of heights skewed towards own tip height. Also
append the genesis block digest for good measure.

Sending the champion digest, if it exists, in a request for a batch of
blocks ensures that these requests can work their way up arbitrarily
long forks. Supplying the genesis block hash, means that the fork can go
all the way back to genesis. In the common case where forks are shallow,
the distribution of heights close to own tip means that there is little
reconciliation work.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>