feat: Add RustyArchivalBlockMmr for schema-access
Previously, we didn't store the schema for the archival-block-mmr,
meaning that we had no way of persisting this data structure to disk
since that functionality is implemented in the `SimpleRustyStorage`
struct which contains a reference to the database. This commit adds a
new data structure `RustyArchivalBlockMmr` which contains the archival
MMR and the schema.

Adding the schema in the form of `SimpleRustyStorage` to the archival
MMR structure in archival state allows us to implement the
`StorageWriter` which allows us to persist the data structure to disk.

Finally, the newly added `persist` trait function is called for
`RustyArchivalBlockMmr` persisting the data to disk.

This construction mirrors the design of `RustyArchivalMutatorSet`.

This closes #345.