fix: cancel proving job if composing stops
addresses #348

Adds plumbing such that when composing task is cancelled the associated
job will get a cancel message and end itself, including the external
proving process, if running.

This first commit relies on tokio's Command::kill_on_drop(true)
functionality, by which the tokio runtime kill's the spawned process in
the background after its handle is dropped (by the proving-job).

This achieves the basic cancellation functionality but it also allows a
window of time during which the old proving process is still running and
a new job begins which starts a second proving process.  Nor does it
detect if the proving process cannot be killed for some reason.

The next commit will address these deficiencies.