refactor: Add `SpendingKey` variant `RawHashLock`
Variant `RawHashLock` is also added to `KeyType`.

The motivation for this change is that the wallet needs to be capable
of spending guesser fee UTXOs. Guesser fee UTXOs are protected by hash
locks whose unlock keys coincide with the nonce preimage. Only the
guesser knows this value.

This spending key type does not have a corresponding address. More
importantly, it breaks the correspondence between spending keys and
receiving addresses, because now there is a spending key type that has
no associated receiving address.

A lot of changes in this commit are fallout from this broken
correspondence. In particular, a lot of methods on `SpendingKey`
assumed a corresponding receiving address and now wrap their return
type in an `Option`, which must be caught gracefully further down the
pipeline.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>