refactor(Block): Require guesser gets at least 1/4 of block subsidy
Require that guesser gets at least 1/4 of block subsidy.

This change incentivizes the rational guesser to know the preimage of
the nonce. If this rule was not there, a mining venture could set the
guesser fraction to 0 leaving the entire block subsidy for the composer.
With this change, the composer could skip on invocation of Tip5 since
the nonce would not have to be the output of Tip5 but could instead be
an arbitrary number. This would give centralized mining ventures and
unfair advantage and thus disincentivize Neptune's two-step mining.

This change also makes it easier for composers to pick up multiple
fee-paying transactions and merge them into an existing block proposal,
thus increasing the guesser reward. If mining venture does not bother
with using a Tip5-digest for the nonce, they will not have an incentive
to pick up and merge fee-paying transactions while PoW guessing is still
ongoing.