fix(peer_loop): Don't grab read lock twice
Fixes a bug where read lock was grabbed twice within the handling of
`PeerMessage::BlockNotification`.

From the documentation of `tokio::sync::rwlock::RwLock`, we see that an
attempt to acquire the read lock twice within the same task can easily
lead to deadlocks, as any request to acquire the write lock will be
given priority, meaning that no read locks will be handed out before the
write lock has been granted and released.

So, the task requesting the write lock will be waiting for the read lock
to be released, and the task attempting to acquire the read lock a 2nd
time will never complete since no read locks are handed out when a write
lock has been requested.

```
The priority policy of Tokio's read-write lock is fair (or write-preferring),
in order to ensure that readers cannot starve writers. Fairness is ensured
using a first-in, first-out queue for the tasks awaiting the lock; if a task
that wishes to acquire the write lock is at the head of the queue, read
locks will not be given out until the write lock has been released. This is
in contrast to the Rust standard library's std::sync::RwLock, where the
priority policy is dependent on the operating system's implementation.
```

This closes #350.

Co-authored-by: dan-da <danda@neptune.cash>