refactor(archival_state): Add block-hash witness to BlockRecord
Add two digests to each record to allow the calculation of the block
hash without knowledge of the entire block kernel.

This emulates a quality of Bitcoin where header-first synchronization
can be used as a PoW-shield against the storing of irrelevant blocks.

Here, we merely use these block hash witnesses to add a `pow_witness`
field to the SyncChallengeResponse. For now, it's only checked that this
field forms a valid chain from the tip. But future modifications can add
a check that these PoW values represent meaningful difficulties.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>