refactor!: Add `guesser_digest` to `BlockHeader`
The guesser fee UTXOs are locked by a hash lock produced from this
digest.

This ensures that hash rate is independent of the composer's guesser-fee
policy/the guesser fee.

Supersedes #355.

We believe that this means that you have to invoke Tip5 exactly seven
times in order to calculate the block hash, and that no shorter path
exists. Prior to this commit, the nonce could be calculated from the
nonce preimage (resulting in eight Tip5 invocations), or the nonce could
be chosen at random (resulting in seven Tip5 invocations). These two
paths to a block hash have now been merged, eliminating a bad incentive
and a discrepency that leads to miner-centralization.

The other parts of the block hash than those coming from the header are
fixed by the block proof, so they'll be **much** more expensive to
modify than the header-nonce.

Implements suggestion from:
https://github.com/Neptune-Crypto/neptune-core/pull/355#issuecomment-2621935483

Co-authored-by: Alan Szepieniec <alan@neptune.cash>