refactor(SyncChallenge): don't check difficulties if own tip is genesis
Avoids a false negative that can happen with very low difficulties when
a newly started node receives a SyncChallengeResponse for a list of
blocks with very low difficulties that has a narrow "difficulty hump" in
the middel. This change is motivated by a false negative on a sync
challenge response observed on a client syncing from genesis on
betanet-v2.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>