refactor!: Add own timestamp to peer handshake
In case we want this value in the future, it's convenient to have the
peer's reported timestamp in the handshake.

Currently, we only use this value to log a warning if peer's reported
time is different from our. But we store it on the `PeerInfo` struct in
case we want to use it to e.g. have a network-adjusted time.