refactor: Make guesser key deterministic
Previously, guesser keys were sampled at random, leading to:
 - complicated machinery to store the keys and use them at the right
   time;
 - impossibility of backing them up in case the wallet files disappear
   or are corrupted.

The guesser keys are now deterministically derived from
 - the wallet secret seed;
 - the raw hash lock flag (being 0);
 - the predecessor block's hash.

This commit applies no other changes. In particular, the wallet logic
for managing guesser keys treats them as though they were random.
Specifically, these keys are still stored in table `guesser_preimages`
of the wallet database even though in theory, all data required to
reproduce them is present without accessing this table. This redundant
storage is not obviously undesirable, as the alternative might incur a
performance penalty.

If the wallet database disappears or is corrupted, the present client
has no functionality to automatically re-generate the guesser keys.
However, the point is that after this change we (or even others) can
write that functionality with retroactive effect when needed.