refactor(`HashLock`): Drop convenience `From` implementations
While `HashLock` wraps a `Digest`, implementing `From<Digest>` for it
and `From<HashLock>` for `Digest` can be confusing because it does not
distinguish whether the preimage or the after-image should be used. To
make this determination, the programmer requires knowledge about
`HashLock`. So the new interface requires less contextual knowledge.

Co-authored-by: Thorkil Værge <thor@neptune.cash>