perf: Add option to maintain empty tx in mempool
This should give a faster composition in some cases, when there are no
other transactions in the mempool. It's likely to slow down composition
if there *are* transactions in the mempool though.

The trick involves keeping an empty (no inputs, no outputs) transaction
in the mempool and keeping this updated with each block. This replaces
the cost of the production of 1 single proof (2^{21} padded height) with
the production of an update (1^{19} padded height). This saves around 27
% of composition time in the case where there are no fee-paying
transactions to include in the block.