refactor(`wallet`): Change determination of "spent" UTXOs
Previously, a `MonitoredUtxo` was counted as spent if its
`spent_in_block` field was set. This field cannot be relied upon because
the block in which the UTXO was spent might have been reorganized away,
and if this happens, that field does not get unset. (We probably want to
fix that in a separate task, or stop using the field altogether.)

The field `spent_in_block` is used for producing a view of history.
UTXOs are shown as spent if this field is set *and* if it has a mutator
set membership proof whose validation fails. Note that this history
screen offers no guarantees of accuracy in the event of (sufficiently
deep) reorganizations, in particular the `spent_in_block` will point to
the wrong block.

This commit exchanges reliance on this field for a validation of the
UTXO's mutator set membership proof against the current mutator set. If
the verification does not go through, then the UTXO is counted as spent.

To accomodate this new check, (a reference to) the mutator set accumulator
is passed to some functions.

Also: struct `WalletStatus` is modified by combining what was previously
two fields: `unsynced_spent` and `unsynced_unspent` to one field, `unsynced`.
With this new way of checking if a monitored UTXO is spent, it has to be
synced to the current tip before we can answer that question.

Fixes #328.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>