refactor(wallet): Use MSMP validity for compiling balance history
With this change, we don't just use the `spent_in_block` field to create
negative balance events, but also check that the MS membership proof
does *not* verify against the current mutator set. This will remove,
from the list of balance events, those UTXOs that have been spent in a
block that was reorganized away.

See previous commit message for a deeper explanation.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>