refactor(RPC): Filter out spent UTXOs smarter
Avoid using the flaky `spent_in_block` field on `MonitoredUtxo` to
filter out spent UTXOs. Instead, test whether the mutator set
membership proof is valid. If it is not valid, the UTXO was spent, and
in this case it should be omitted from the list.

The interface to `get_all_own_coins_with_possible_timelocks` is modified
accordingly -- in particular, to decide this boolean, both the tip hash
and the mutator set accumulator are required.

Cf. #368.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>