style(`NativeCurrencyAmount`): Drop ambiguous `FromStr` conversion
After the name change `NeptuneCoins` -> `NativeCurrencyAmount`
(a2038ef), the method `from_str` is ambiguous: it obviously takes a
string, but does this string represent an amount of nau or an amount
of coins?

This commit disambiguates by renaming the method to `coins_from_str`
and putting on `NativeCurrencyAmount` proper as opposed to an `impl`
of `FromStr`.

Fixes #324 more.