fix: remove dashboard balance discrepancy
fixes #346

This commit refactors how balances are calculated and displayed. It
normalizes balance related functions to use the terminology:
 confirmed/unconfirmed and available/total.

Balances are displayed in the dashboard overview as 4 fields in 2 rows:
 confirmed balance:    total, available
 unconfirmed balance:  total, available

The unconfirmed balance row is only displayed if either of the two
fields differs from its counterpart in the confirmed balance row.

In neptune-core, the WalletStatus type is refactored a bit to remove
unused fields and methods.  The unsynced field is only used in tests,
so is now decorated with #[cfg(test]).

In WalletState, the balance getting functions now take &WalletState
which avoids duplicate work (iterating over all monitored utxos) by
callers that call multiple of these functions, in particular the
dashboard_overview_data() RPC.

neptune-core changes:
 * WalletState::confirmed_balance() --> confirmed_available_balance()
 * add WalletState::confirmed_total_balance()
 * WalletState::unconfirmed_balance() --> unconfirmed_available_balance()
 * add WalletState::unconfirmed_total_balance()
 * walletState::get_wallet_status_from_lock() --> get_wallet_status()
 * added a big doc-comment describing WalletStatus
 * WalletStatus::unsynced is now #[cfg(test)]
 * WalletStatus::synced_unspent_liquid_amount() -> synced_unspent_available_amount()
 * add WalletStatus::synced_unspent_total_amount()
 * remove unused WalletStatus::unsynced_amount()
 * remove unused WalletStatus::synced_spent_amount()
 * remove unused impl Display for WalletStatus
 * rpc confirmation_Internal() is now 1 based instead of 0 based.
     because block that mines tx is 1st confirmation.
 * update tests

dashboard changes:
 * overview screen displays 4 balance fields instead of 3 but in a more
   compact and readable representation.
 * balance fields have a width of 17 digits.
    (8 whole coin, 1 decimal point, 8 decimals)
 * "(x confirmations)" --> "(1 confirmation)" (plural aware)
 * "(x confirmations) now starts at 1. block that mines tx is 1st confirmation.
 * only display unconfirmed balance row if needed.

neptune-cli changes:
 * command synced-balance is now confirmed-available-balance
 * command synced-balance-unconfirmed is now
   unconfirmed-available-balance

(for consistency with naming throughout codebase)